refactor(core): normalize angle constants to float across core and scripts
- replace implicit M_PI and follow-angle conversions with explicit float casts
- normalize angle handling in core systems, AI, movement, spells and scripts
- reduce mixed float/double usage in trigonometric and orientation code paths

+++ src/server/game/AI/ScriptedAI/ScriptedCreature.cpp
@@ -1,9 +1,19 @@
-/* Copyright (C) 2008-2012 TrinityCore <http://www.trinitycore.org/>
+/*
+ * This file is part of the DestinyCore Project. See AUTHORS file for Copyright information
  *
- * Thanks to the original authors: ScriptDev2 <https://scriptdev2.svn.sourceforge.net/>
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
  *
- * This program is free software licensed under GPL version 2
- * Please see the included DOCS/LICENSE.TXT for more information */
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
 
 #include "BrawlersGuild.h"
 #include "Cell.h"
@@ -1165,7 +1175,7 @@ void GetPosInRadiusWithRandomOrientation(Unit* unit, float dist, float &x, float
 
 void GetRandPosFromCenterInDist(float centerX, float centerY, float dist, float& x, float& y)
 {
-    float randOrientation = frand(0, 2 * M_PI);
+    float randOrientation = frand(0, float(2 * M_PI));
 
     x = centerX + (dist * cos(randOrientation));
     y = centerY + (dist * sin(randOrientation));
