+++ src/server/scripts/Draenor/Highmaul/boss_brackenspore.cpp
@@ -1,10 +1,19 @@
-////////////////////////////////////////////////////////////////////////////////
-//
-//  MILLENIUM-STUDIO
-//  Copyright 2016 Millenium-studio SARL
-//  All Rights Reserved.
-//
-////////////////////////////////////////////////////////////////////////////////
+/*
+ * This file is part of the DestinyCore Project. See AUTHORS file for Copyright information
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
 
 #include "ScriptMgr.h"
 #include "highmaul.hpp"
@@ -560,12 +569,12 @@ struct boss_brackenspore : public BossAI
         if (map == nullptr)
             return;
 
-        float o = frand(0.0f, 2 * M_PI);
+        float o = frand(0.0f, float(2 * M_PI));
 
         /// Use different spawn radius depending on orientation
         float radius = GetSpawnRangeByOrientation(o);
 
-        float oStep = 2 * M_PI / 30.0f;
+        float oStep = float(2 * M_PI / 30.0f);
         float x = beachCenter.x + (radius * cos(o));
         float y = beachCenter.y + (radius * sin(o));
         float z = map->GetHeight(x, y, MAX_HEIGHT);
